code/requirements.txt
discord
asyncio
load_dotenv
json5

